Refactored getUserInfos in order to retrieve roles
This function previously returned a manually formatted string colon-separated.
This may be rapidily handled for a two-fields composite element, but this
may become quickly painful for adding or changing fields types.

For this, a UserInfos was build as a model structure, and is returned
from the /info route. This is the one being saved into the Redis session
cache.